\name{leastFavorableRadius-methods}
\docType{methods}
\alias{leastFavorableRadius-methods}
\alias{leastFavorableRadius,L2RegTypeFamily,Neighborhood,asGRisk-method}

\title{ Methods for Function leastFavorableRadius in Package `ROptRegTS' }
\description{
  Methods for function \code{leastFavorableRadius} in package \pkg{ROptRegTS}.
}
\section{Methods}{
\describe{
  \item{L2Fam = "L2RegTypeFamily", neighbor = "Neighborhood", risk = "asGRisk"}{ 
    The least favorable radius and the corresponding inefficiency 
    are computed. }
}}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
\seealso{\code{\link[ROptEstOld]{leastFavorableRadius}}}
\keyword{methods}
\keyword{robust}
