\name{radiusMinimaxIC-methods}
\docType{methods}
\alias{radiusMinimaxIC-methods}
\alias{radiusMinimaxIC,L2RegTypeFamily,Neighborhood,asGRisk-method}

\title{ Methods for Function radiusMinimaxIC in Package `ROptRegTS' }
\description{
  Methods for function \code{radiusMinimaxIC} in package \pkg{ROptRegTS}.
}
\section{Methods}{
\describe{
  \item{L2Fam = "L2RegTypeFamily", neighbor = "Neighborhood", risk = "asGRisk"}{ 
    computation of the radius minimax IC for an L2 differentiable 
    regression-type family. }
}}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
\seealso{\code{\link[ROptEstOld]{radiusMinimaxIC}}}
\keyword{methods}
\keyword{robust}
